requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
typer==0.9.0
litellm==1.40.0
qdrant-client==1.7.0
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3